Swerve
Swerve is the first video game I ever created, and one of the first coding projects I ever started. (Do I get a pass for having almost no comments?)
To actually play the game, you will need to install PyGame as well as download the folder with all the assets. However, if you just want to see it run, I have posted a video of me playing the game.